backend/requirements.txt
fastapi==0.135.3
uvicorn==0.34.0
pandas==3.0.2
python-multipart==0.0.20
python-dotenv==1.0.1